+++ classes/privacy/provider.php
@@ -15,18 +15,33 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Auth plugin "LDAP SyncPlus" - Version file
+ * Auth plugin "LDAP SyncPlus" - Privacy provider
  *
  * @package    auth_ldap_syncplus
- * @copyright  2014 Alexander Bias, Ulm University <alexander.bias@uni-ulm.de>
+ * @copyright  2018 Alexander Bias, Ulm University <alexander.bias@uni-ulm.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
+namespace auth_ldap_syncplus\privacy;
+
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->component = 'auth_ldap_syncplus';
-$plugin->version = 2018051600;
-$plugin->release = 'v3.4-r3';
-$plugin->requires = 2017111300;
-$plugin->maturity = MATURITY_STABLE;
-$plugin->dependencies = array('auth_ldap' => 2017051500);
+/**
+ * Privacy Subsystem implementing null_provider.
+ *
+ * @package    auth_ldap_syncplus
+ * @copyright  2018 Alexander Bias, Ulm University <alexander.bias@uni-ulm.de>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class provider implements \core_privacy\local\metadata\null_provider {
+
+    /**
+     * Get the language string identifier with the component's language
+     * file to explain why this plugin stores no data.
+     *
+     * @return string
+     */
+    public static function get_reason() : string {
+        return 'privacy:metadata';
+    }
+}